test/dependency_source: remove a flaky assertion
The assertion here is logically sound, but would commonly flake
for PEBKAC reasons: users in a development environment might
forget to `pip install .` between version bumps, causing the
tests to see one `__version__` and the running code to see another.

The assertion was a dupe of a much less flaky assertion, so removing
it causes no coverage change.